refactor(test): lift the watching-controller factory out of its suite
`WatchingControllerTests.swift` was at 595 of its 600-line cap, so the file
refused any new case at all. The controller factory and `CallCounter` are not
specific to that suite, and a second suite now needs the same wiring, so both
move to their own files and the suite keeps a thin delegator that supplies its
own `tmpDir`.

Pure move: no test changes behaviour, and the suite drops to 566 lines, which
leaves room for the next case rather than putting the next author back at the
cap. Same remedy the `AsyncGate` extraction used.
